migrate Cupcake example to Gradle 9.5, AGP 9.2, Gradle Version catalog and new project structure
