+++ examples/nav_cupcake/androidApp/src/main/AndroidManifest.xml
