--- examples/nav_cupcake/composeApp/src/commonMain/kotlin/org/jetbrains/nav_cupcake/ui/components/CommonUi.kt
@@ -1,36 +1,0 @@
-/*
- * Copyright (C) 2023 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jetbrains.nav_cupcake.ui.components
-
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import cupcake.composeapp.generated.resources.Res
-import cupcake.composeapp.generated.resources.subtotal_price
-import org.jetbrains.compose.resources.stringResource
-
-/**
- * Composable that displays formatted [price] that will be formatted and displayed on screen
- */
-@Composable
-fun FormattedPriceLabel(subtotal: String, modifier: Modifier = Modifier) {
-    Text(
-        text = stringResource(Res.string.subtotal_price, subtotal),
-        modifier = modifier,
-        style = MaterialTheme.typography.headlineSmall
-    )
-}
