--- examples/nav_cupcake/composeApp/src/desktopMain/kotlin/main.kt
@@ -1,9 +1,0 @@
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.window.Window
-import androidx.compose.ui.window.application
-
-fun main() = application {
-    Window(onCloseRequest = ::exitApplication, title = "Cupcake") {
-        App()
-    }
-}
