+++ examples/nav_cupcake/desktopApp/rules.pro
