+++ examples/nav_cupcake/desktopApp/src/jvmMain/kotlin/main.kt
@@ -1,0 +1,8 @@
+import androidx.compose.ui.window.Window
+import androidx.compose.ui.window.application
+
+fun main() = application {
+    Window(onCloseRequest = ::exitApplication, title = "Cupcake") {
+        App()
+    }
+}
