+++ examples/nav_cupcake/gradle.properties
@@ -1,16 +1,7 @@
 kotlin.code.style=official
-
-#Gradle
-org.gradle.jvmargs=-Xmx8G -Dfile.encoding=UTF-8 -Dkotlin.daemon.jvm.options\="-Xmx8G"
-
-
-#Android
-android.nonTransitiveRClass=true
 android.useAndroidX=true
-
-#MPP
-kotlin.mpp.androidSourceSetLayoutVersion=2
-kotlin.mpp.enableCInteropCommonization=true
-
-#Development
-development=true
+org.gradle.jvmargs=-Xmx8g
+org.gradle.caching=true
+kotlin.version=2.3.20
+agp.version=9.2.1
+compose.version=1.10.1
