+++ examples/nav_cupcake/settings.gradle.kts
@@ -1,12 +1,11 @@
-rootProject.name = "Cupcake"
+@file:Suppress("UnstableApiUsage")
 enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
 
 pluginManagement {
     repositories {
-        google()
         gradlePluginPortal()
-        mavenCentral()
         maven("https://packages.jetbrains.team/maven/p/cmp/dev")
+        google()
     }
 }
 
@@ -16,6 +15,23 @@ dependencyResolutionManagement {
         mavenCentral()
         maven("https://packages.jetbrains.team/maven/p/cmp/dev")
     }
+
+    versionCatalogs {
+        create("libs") {
+            version("kotlin", extra["kotlin.version"].toString())
+            version("compose", extra["compose.version"].toString())
+            version("agp", extra["agp.version"].toString())
+        }
+    }
 }
 
-include(":composeApp")
+plugins {
+    id("org.gradle.toolchains.foojay-resolver-convention") version("1.0.0")
+}
+
+rootProject.name = "Cupcake"
+
+include(":shared")
+include(":androidApp")
+include(":desktopApp")
+include(":webApp")
