+++ examples/nav_cupcake/shared/src/commonMain/composeResources/drawable/cupcake.xml
