+++ examples/nav_cupcake/shared/src/commonMain/composeResources/values/strings.xml
