+++ examples/nav_cupcake/shared/src/commonMain/kotlin/org/jetbrains/nav_cupcake/data/OrderUiState.kt
