+++ examples/nav_cupcake/shared/src/commonMain/kotlin/org/jetbrains/nav_cupcake/ui/SelectOptionScreen.kt
@@ -1,0 +1,133 @@
+/*
+ * Copyright (C) 2023 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.jetbrains.nav_cupcake.ui
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.selection.selectable
+import androidx.compose.material3.*
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import org.jetbrains.nav_cupcake.ui.components.FormattedPriceLabel
+import org.jetbrains.nav_cupcake.ui.theme.CupcakeTheme
+import cupcake.shared.generated.resources.*
+import org.jetbrains.compose.resources.stringResource
+
+/**
+ * Composable that displays the list of items as [RadioButton] options,
+ * [onSelectionChanged] lambda that notifies the parent composable when a new value is selected,
+ * [onCancelButtonClicked] lambda that cancels the order when user clicks cancel and
+ * [onNextButtonClicked] lambda that triggers the navigation to next screen
+ */
+@Composable
+fun SelectOptionScreen(
+    subtotal: String,
+    options: List<String>,
+    onSelectionChanged: (String) -> Unit = {},
+    onCancelButtonClicked: () -> Unit = {},
+    onNextButtonClicked: () -> Unit = {},
+    modifier: Modifier = Modifier
+) {
+    var selectedValue by rememberSaveable { mutableStateOf("") }
+
+    Column(
+        modifier = modifier,
+        verticalArrangement = Arrangement.SpaceBetween
+    ) {
+        Column(modifier = Modifier.padding(16.dp)) {
+            options.forEach { item ->
+                Row(
+                    modifier = Modifier.selectable(
+                        selected = selectedValue == item,
+                        onClick = {
+                            selectedValue = item
+                            onSelectionChanged(item)
+                        }
+                    ),
+                    verticalAlignment = Alignment.CenterVertically
+                ) {
+                    RadioButton(
+                        selected = selectedValue == item,
+                        onClick = {
+                            selectedValue = item
+                            onSelectionChanged(item)
+                        }
+                    )
+                    Text(item)
+                }
+            }
+            HorizontalDivider(
+                thickness = 1.dp,
+                modifier = Modifier.padding(bottom = 16.dp)
+            )
+            FormattedPriceLabel(
+                subtotal = subtotal,
+                modifier = Modifier
+                    .align(Alignment.End)
+                    .padding(
+                        top = 16.dp,
+                        bottom = 16.dp
+                    )
+            )
+        }
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(16.dp),
+            horizontalArrangement = Arrangement.spacedBy(16.dp),
+            verticalAlignment = Alignment.Bottom
+        ) {
+            OutlinedButton(
+                modifier = Modifier.weight(1f),
+                onClick = onCancelButtonClicked
+            ) {
+                Text(stringResource(Res.string.cancel))
+            }
+            Button(
+                modifier = Modifier.weight(1f),
+                // the button is enabled when the user makes a selection
+                enabled = selectedValue.isNotEmpty(),
+                onClick = onNextButtonClicked
+            ) {
+                Text(stringResource(Res.string.next))
+            }
+        }
+    }
+
+}
+
+@Preview
+@Composable
+fun SelectOptionPreview() {
+    CupcakeTheme {
+        SelectOptionScreen(
+            subtotal = "299.99",
+            options = listOf("Option 1", "Option 2", "Option 3", "Option 4"),
+            modifier = Modifier.fillMaxHeight()
+        )
+    }
+}
