+++ examples/nav_cupcake/webApp/src/wasmJsMain/resources/index.html
